notification_system/orchestrator/requirements.txt
Flask
confluent-kafka
pydantic
supabase
Flask-Cors
python-dotenv
Werkzeug